test: RepositoryFake 삭제
