+++ src/main/java/roomescape/infrastructure/dao/MemberDao.java
@@ -1,70 +1,18 @@
 package roomescape.infrastructure.dao;
 
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 import roomescape.domain.model.Member;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
-@Repository
-public class MemberDao {
+public interface MemberDao {
 
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert simpleJdbcInsert;
+    Long save(Member member);
 
-    private static final RowMapper<Member> ROW_MAPPER = (resultSet, rowNum) ->
-            new Member(
-                    resultSet.getLong("id"),
-                    resultSet.getString("name"),
-                    resultSet.getString("email"),
-                    resultSet.getString("password"),
-                    resultSet.getString("role")
-            );
+    boolean existByEmail(final String email);
 
-    public MemberDao(final JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("member")
-                .usingGeneratedKeyColumns("id");
-    }
+    Member findByEmail(String email);
 
-    public Long save(Member member) {
-        Map<String, Object> parameters = new HashMap<>();
-        parameters.put("name", member.getName());
-        parameters.put("email", member.getEmail());
-        parameters.put("password", member.getPassword());
-        parameters.put("role", member.getRole());
+    Member findById(final Long memberId);
 
-        return simpleJdbcInsert.executeAndReturnKey(parameters).longValue();
-    }
-
-    public boolean existByEmail(final String email) {
-        String query = "SELECT EXISTS (SELECT 1 FROM member WHERE email = ?)";
-        return jdbcTemplate.queryForObject(query, Boolean.class, email);
-    }
-
-    public Member findByEmail(String email) {
-        String query = "SELECT * FROM member WHERE email = ?";
-        return jdbcTemplate.queryForObject(
-                query,
-                ROW_MAPPER,
-                email);
-    }
-
-    public Member findById(final Long memberId) {
-        String query = "SELECT * FROM member WHERE id = ?";
-        return jdbcTemplate.queryForObject(
-                query,
-                ROW_MAPPER,
-                memberId);
-    }
-
-    public List<Member> findAll() {
-        String query = "SELECT * FROM member";
-        return jdbcTemplate.query(query, ROW_MAPPER);
-    }
+    List<Member> findAll();
 }
