+++ src/main/java/roomescape/infrastructure/dao/ThemeDao.java
@@ -1,77 +1,20 @@
 package roomescape.infrastructure.dao;
 
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 import roomescape.domain.model.Theme;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
-@Repository
-public class ThemeDao {
+public interface ThemeDao {
 
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert simpleJdbcInsert;
-    private static final RowMapper<Theme> ROW_MAPPER = (resultSet, rowNum) -> new Theme(
-            resultSet.getLong("id"),
-            resultSet.getString("name"),
-            resultSet.getString("description"),
-            resultSet.getString("thumbnail")
-    );
+    List<Theme> findAll();
 
-    public ThemeDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("theme")
-                .usingGeneratedKeyColumns("id");
-    }
+    Theme save(Theme theme);
 
-    public List<Theme> findAll() {
-        String query = "SELECT * FROM theme";
-        return jdbcTemplate.query(query, ROW_MAPPER);
-    }
+    int deleteById(Long id);
 
-    public Theme save(Theme theme) {
-        Map<String, Object> parameters = new HashMap<>();
-        parameters.put("name", theme.getName());
-        parameters.put("description", theme.getDescription());
-        parameters.put("thumbnail", theme.getThumbnail());
+    Theme findById(Long id);
 
-        long id = simpleJdbcInsert.executeAndReturnKeyHolder(parameters).getKey().longValue();
-        return new Theme(id, theme.getName(), theme.getDescription(), theme.getThumbnail());
-    }
+    boolean existByName(String name);
 
-    public int deleteById(Long id) {
-        String query = "DELETE FROM theme WHERE id = ?";
-        return jdbcTemplate.update(query, id);
-    }
-
-    public Theme findById(Long id) {
-        String sql = "SELECT * from theme where id = ?";
-        return jdbcTemplate.queryForObject(sql, ROW_MAPPER, id);
-    }
-
-    public boolean existByName(String name) {
-        String sql = "SELECT EXISTS (SELECT 1 from theme where name = ?)";
-        return jdbcTemplate.queryForObject(sql, Boolean.class, name);
-    }
-
-    public List<Theme> findPopular(int count) {
-        String sql = """
-                SELECT count(*), t.id, t.name, t.description, t.thumbnail FROM theme AS t
-                LEFT JOIN (
-                    SELECT * 
-                    FROM reservation AS r 
-                    WHERE PARSEDATETIME(r.date, 'yyyy-MM-dd') BETWEEN TIMESTAMPADD(DAY, -7, CURRENT_DATE) AND TIMESTAMPADD(DAY, -1, CURRENT_DATE)
-                ) AS r ON t.id = r.theme_id
-                GROUP BY t.id
-                ORDER BY count(*) DESC
-                limit ?
-                """;
-        return jdbcTemplate.query(sql, ROW_MAPPER, count);
-    }
-
+    List<Theme> findPopular(int count);
 }
