+++ src/test/java/roomescape/fake/ReservationDaoFake.java
@@ -1,6 +1,7 @@
 package roomescape.fake;
 
 import roomescape.domain.model.Reservation;
+import roomescape.infrastructure.dao.ReservationDao;
 
 import java.time.LocalDate;
 import java.util.HashMap;
@@ -8,11 +9,12 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 
-public class FakeReservationDao {
+public class ReservationDaoFake implements ReservationDao {
 
     private static final AtomicLong IDX = new AtomicLong();
     private static final Map<Long, Reservation> RESERVATIONS = new HashMap<>();
 
+    @Override
     public Reservation save(Reservation reservation) {
         Long id = IDX.getAndIncrement();
         Reservation newReservation = new Reservation(id, reservation.getMemberId(), reservation.getDate(), reservation.getTime(), reservation.getTheme());
@@ -20,10 +22,12 @@ public class FakeReservationDao {
         return newReservation;
     }
 
+    @Override
     public List<Reservation> findAll() {
         return RESERVATIONS.values().stream().toList();
     }
 
+    @Override
     public List<Reservation> findByThemeIdAndMemberIdAndDate(final Long themeId, final Long memberId, final LocalDate dateFrom, final LocalDate dateTo) {
         return RESERVATIONS.values().stream()
                 .filter(reservation -> {
@@ -35,6 +39,7 @@ public class FakeReservationDao {
                 }).toList();
     }
 
+    @Override
     public int deleteById(Long id) {
         if (RESERVATIONS.containsKey(id)) {
             RESERVATIONS.remove(id);
@@ -43,6 +48,7 @@ public class FakeReservationDao {
         return 0;
     }
 
+    @Override
     public boolean existByTimeId(Long timeId) {
         long count = RESERVATIONS.entrySet().stream()
                 .filter(e -> e.getValue().getTime().getId().equals(timeId))
@@ -50,6 +56,7 @@ public class FakeReservationDao {
         return count != 0;
     }
 
+    @Override
     public boolean existByThemeId(Long themeId) {
         long count = RESERVATIONS.entrySet().stream()
                 .filter(e -> e.getValue().getTheme().getId().equals(themeId))
@@ -57,6 +64,7 @@ public class FakeReservationDao {
         return count != 0;
     }
 
+    @Override
     public boolean existByTimeIdAndThemeIdAndDate(Long timeId, Long themeId, LocalDate date) {
         long count = RESERVATIONS.entrySet().stream()
                 .filter(e -> e.getValue().getTime().getId().equals(timeId)
@@ -66,6 +74,7 @@ public class FakeReservationDao {
         return count != 0;
     }
 
+    @Override
     public List<Long> findBookedTimes(final Long themeId, final LocalDate date) {
         return RESERVATIONS.values().stream()
                 .filter(reservation -> reservation.getTheme().getId().equals(themeId) && reservation.getDate().equals(date))
