+++ src/test/java/roomescape/fake/ThemeDaoFake.java
@@ -1,25 +1,24 @@
 package roomescape.fake;
 
-import roomescape.domain.model.Reservation;
 import roomescape.domain.model.Theme;
+import roomescape.infrastructure.dao.ThemeDao;
 
-import java.time.LocalDate;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
-import java.util.stream.Collectors;
 
-public class FakeThemeDao {
+public class ThemeDaoFake implements ThemeDao {
 
     private static final AtomicLong IDX = new AtomicLong();
     private static final Map<Long, Theme> THEMES = new HashMap<>();
 
+    @Override
     public Theme findById(Long id) {
         return THEMES.get(id);
     }
 
+    @Override
     public Theme save(Theme theme) {
         Long id = IDX.getAndIncrement();
         Theme newTheme = new Theme(id, theme.getName(), theme.getDescription(), theme.getThumbnail());
@@ -27,10 +26,12 @@ public class FakeThemeDao {
         return newTheme;
     }
 
+    @Override
     public List<Theme> findAll() {
         return THEMES.values().stream().toList();
     }
 
+    @Override
     public int deleteById(Long id) {
         if (THEMES.containsKey(id)) {
             THEMES.remove(id);
@@ -39,6 +40,7 @@ public class FakeThemeDao {
         return 0;
     }
 
+    @Override
     public boolean existByName(String name) {
         long count = THEMES.values().stream()
                 .filter(theme -> theme.getName().equals(name))
@@ -46,20 +48,9 @@ public class FakeThemeDao {
         return count != 0;
     }
 
-    public List<Theme> findPopular(List<Reservation> reservations, int count) {
-        LocalDate today = LocalDate.now();
-        LocalDate startDate = today.minusDays(7);
-        LocalDate endDate = today.minusDays(1);
-
-        Map<Theme, Long> reservationCount = reservations.stream()
-                .filter(r -> !r.getDate().isBefore(startDate) && !r.getDate().isAfter(endDate))
-                .collect(Collectors.groupingBy(Reservation::getTheme, Collectors.counting()));
-
-        return reservationCount.entrySet().stream()
-                .sorted(Comparator.<Map.Entry<Theme, Long>>comparingLong(Map.Entry::getValue).reversed())
-                .limit(count)
-                .map(Map.Entry::getKey)
-                .toList();
+    @Override
+    public List<Theme> findPopular(final int count) {
+        return List.of();
     }
 
     public void clear() {
